musicplatform/pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["Ammar Mahmoud <amar.link8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.3"
django-rest-knox = "^5.0.1"
cryptography = "^43.0.0"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
